feat: Implement dynamic dashboard with Supabase data
This commit replaces the hardcoded mock data in your dashboard with live data fetched from a Supabase backend.

Key changes include:
- Added a new database migration to create `profiles`, `applications`, and `documents` tables.
- Refactored the dashboard page into a Next.js server component to fetch data directly from Supabase.
- Separated the interactive parts of the dashboard into a new client component (`dashboard-client.tsx`) to handle UI state and interactivity.
- The dashboard now displays your name, your job applications, and your documents from the database.
- Added authentication check to the dashboard page, redirecting unauthenticated users to the login page.

Note: The database migration in `supabase/migrations/01_dashboard_schema.sql` must be applied to the Supabase project for the dashboard to work correctly. No existing testing framework was found, so no new tests were added as part of this change.
